Move everything out of .forkmydots.
After reviewing some existing dotfiles projects, I have found that one common
theme is NOT to clobber any files in the user's home directory at all, but
instead to install dotfiles from the directory resulting from a clone of the
project. This makes sense to me, so I will try it out.